Update index.html
Update docs to explain that some superuser access is likely to be required in most installation scenarios, as highlighted by https://github.com/mobile-shell/mosh/issues/188 .  This is due to the dependency of binaries on dynamically-linked shared libraries, which are unlikely to be all pre-installed on most platforms, and will require superuser access to be installed.  That is until/unless there is a statically-linked option for most server/client on some platforms.